backend/requirements.txt
fastapi
uvicorn
rembg
python-multipart
pillow
mediapipe
opencv-python
